requirements.txt
gunicorn
django
ipython
ipdb
uvicorn
fastapi
python-socketio
rpdb
django-rest-framework
drf-yasg
